Add files via upload
Here is the mouse skin dataset for our paper with title “Rank-based scatter matrix estimation for high dimensional microbiome data". The dataset originates from the paper "Genome-wide mapping of gene-microbiota interactions in susceptibility to autoimmune skin blistering". We would like to point out that the data in "Taxonomy_131_OTU_CMM.txt" is transformed by the operator lg. And the sum of each column is not equal to one because they only contain about 80% of the total sequence.